topic: before responsive

+++ src/com/hello/client/activities/NormalAppActivityMapper.java
@@ -1,5 +1,6 @@
 package com.hello.client.activities;
 
+import com.google.gwt.core.client.GWT;
 import com.google.gwt.place.shared.Place;
 import com.hello.client.activities.addupdate2.AddUpdateContactActivity2;
 import com.hello.client.activities.addupdate2.AddUpdateContactPlace2;
@@ -11,6 +12,7 @@ import com.hello.client.activities.home.HomeActivity;
 import com.hello.client.activities.home.HomePlace;
 import com.hello.client.activities.homepage.HomepageActivity;
 import com.hello.client.activities.homepage.HomepagePlace;
+import com.hello.shared.cache.ContactInfoCache;
 
 
 public class NormalAppActivityMapper implements AsyncActivityMapper {
@@ -23,9 +25,6 @@ public class NormalAppActivityMapper implements AsyncActivityMapper {
 
 	@Override
 	public void getActivity(final Place place, final ActivityCallbackHandler activityCallbackHandler) {
-		HomepageActivity homepageActivity = clientFactory.getHomepageActivity();
-		Runnable loadDataCallback = homepageActivity != null ? homepageActivity::loadData : null;
-		
 		
 		if (place instanceof HomePlace) {
 			activityCallbackHandler.onRecieveActivity( new HomeActivity(clientFactory, place));
@@ -34,36 +33,52 @@ public class NormalAppActivityMapper implements AsyncActivityMapper {
 			activityCallbackHandler.onRecieveActivity( new ContactActivity(clientFactory, place));
 		} 
 		else if (place instanceof HomepagePlace) {
-			activityCallbackHandler.onRecieveActivity(new HomepageActivity(clientFactory, place));
+			HomepageActivity homepageActivity = clientFactory.getHomepageActivity();
+
+		    if (homepageActivity == null) {
+		        homepageActivity = new HomepageActivity(clientFactory, place);
+		        clientFactory.setHomePageActivity(homepageActivity);
+		    } 
+//		    homepageActivity.setPlace(place); // Nếu bạn cần cập nhật place mới
+//	        homepageActivity.loadData();  // xóa đi để trogn homepage load thì sẽ ko bị lặp 2 lần nữa 
+
+		    activityCallbackHandler.onRecieveActivity(homepageActivity);
+		
 		}
 		else if (place instanceof AddUpdateContactPlace2) {
-			
-		    AddUpdateContactPlace2 myPlace = (AddUpdateContactPlace2) place;
-
-		    activityCallbackHandler.onRecieveActivity(
-		        new AddUpdateContactActivity2(
-		            clientFactory,
-		            place,
-		            clientFactory.getGreetingService(), // Lấy service
-		            clientFactory.getDataProvider(), // ListDataProvider<ContactInfo>
-		            myPlace.getActionType(),
-		            loadDataCallback
-		            // actionType (ADD / UPDATE)
-		            // Hoặc null nếu chưa cần callback
-		        )
-		    );
-		}
+            AddUpdateContactPlace2 addUpdatePlace = (AddUpdateContactPlace2) place;
+            AddUpdateContactActivity2 addUpdateActivity = clientFactory.getAddUpdateContactActivity2();
+            if (addUpdateActivity == null) {
+                addUpdateActivity = new AddUpdateContactActivity2(
+                    clientFactory, place, addUpdatePlace.getActionType());
+                clientFactory.setAddUpdateContactActivity2(addUpdateActivity);
+                GWT.log("Created new AddUpdateContactActivity2 instance");
+            } else {
+                // Cập nhật dữ liệu cho instance hiện có
+            	addUpdateActivity.setPlace(place);
+                addUpdateActivity.setActionType(addUpdatePlace.getActionType());
+                String phone = addUpdatePlace.getPhoneNumber();
+                addUpdateActivity.setSelectedContact((phone != null) ? ContactInfoCache.getByPhoneNumber(phone) : null);
+                addUpdateActivity.loadData();
+                GWT.log("Reusing AddUpdateContactActivity2 instance");
+            }
+            activityCallbackHandler.onRecieveActivity(addUpdateActivity);
+        }
 		else if (place instanceof DeleteContactPlace2) {
-			DeleteContactPlace2 myPlace = (DeleteContactPlace2) place;
-
-		    activityCallbackHandler.onRecieveActivity(
-		        new DeleteContactActivity2(
-		                clientFactory,
-		                (DeleteContactPlace2) place,
-		                clientFactory.getGreetingService(),
-		                clientFactory.getDataProvider(),
-		                loadDataCallback
-		            ));
-		}
+            DeleteContactPlace2 deletePlace = (DeleteContactPlace2) place;
+            DeleteContactActivity2 deleteActivity = clientFactory.getDeleteContactActivity2();
+            if (deleteActivity == null) {
+                deleteActivity = new DeleteContactActivity2(
+                    clientFactory, deletePlace);
+                clientFactory.setDeleteContactActivity2(deleteActivity);
+                GWT.log("Created new DeleteContactActivity2 instance");
+            } else {
+                // Cập nhật dữ liệu cho instance hiện có
+                deleteActivity.setSelectedPhoneNumbers(deletePlace.getPhoneNumbers()) ;
+                deleteActivity.start(null, null); // Gọi start để cập nhật UI
+                GWT.log("Reusing DeleteContactActivity2 instance");
+            }
+            activityCallbackHandler.onRecieveActivity(deleteActivity);
+        }
 	}
 }
